requirements.txt
numpy==1.24.4
Pillow==10.0.0
tensorflow==2.18.0
